v2/gas_sponsorship/python/pyproject.toml
[project]
name = "gas_sponsorship_python_example"
version = "1.0.0"
description = "Example: Gas Sponsorship for Earn Account Funding and Position Management using Compass API"
dependencies = [
    "compass-api-sdk",
    "python-dotenv",
    "web3",
    "eth-account",
]

